Remove stderr hack, write choosen path to file
Until now ncurses used stderr. This enabled us to write the actual
result path to stdout and the 'cd' in the shell function got it
properly.

This was a hack since it misused the output streams. Because of this
there was also no proper presentation of error messages to the user.
More fatal this hack broke opening the editor. Vim reported "Warning:
Output is not to a terminal".

This is fixed by writing the result path into "$HOME/goto.result". The
default path is ".".